Corrección de vulnerabilidad
Se corrige el que se permitiese inyectar código en varias de las llamadas a AutoFirma en Windows y macOS de tal forma que pudiesen ejecutarse scripts y aplicaciones en el equipo del usuario. En Windows ocurría cuando se usaba la función getShort() para calcular el nombre corto de las rutas de las bibliotecas PKCS#11. En macOS ocurría al pasar el id del proceso lanzador de AutoFirma para que se cerrase al finalizar la ejecución de AutoFirma.

+++ afirma-simple/src/main/java/es/gob/afirma/standalone/so/macos/MacUtils.java
@@ -88,38 +88,45 @@ public class MacUtils {
 		}
 	}
 
-
     /** Ejecuta un fichero de scripts.
-	 * @param scriptFile Fichero de <i>script</i>.
-	 * @param administratorMode <code>true</code> el <i>script</i> se ejecuta con permisos de adminsitrador, <code>false</code> en caso contrario.
-	 * @param delete <code>true</code> se borra el fichero despu&eacute;s de haberse ejecutado.
-	 * @return El objeto que da como resultado el <i>script</i>.
-	 * @throws IOException Excepci&oacute;n lanzada en caso de ocurrir alg&uacute;n error en la ejecuci&oacute;n del <i>script</i>.
-     * @throws InterruptedException Cuando el proceso se ve interrumpido (posiblemente por el usuario). */
-	public static Object executeScriptFile(final File scriptFile, final boolean administratorMode, final boolean delete) throws IOException, InterruptedException {
+	 * @param scriptFile Ruta del fichero de <i>script</i>.
+	 * @param administratorMode {@code true} el <i>script</i> se ejecuta con permisos de adminsitrador,
+	 * {@code false} en caso contrario.
+	 * @param delete {@code true} borra el fichero despu&eacute;s de haberse ejecutado, {@code false} no hace nada.
+	 * @return La cadena que da como resultado el <i>script</i>.
+	 * @throws IOException Cuando ocurre un error en la ejecuci&oacute;n del <i>script</i>.
+     * @throws InterruptedException  Cuando se interrumpe la ejecuci&oacute;n del script (posiblemente por el usuario). */
+	public static String executeScriptFile(final File scriptFile, final boolean administratorMode, final boolean delete) throws IOException, InterruptedException {
+
+		LOGGER.info("Se ejecuta el script: " + scriptFile); //$NON-NLS-1$
 
 		final AppleScript script = new AppleScript(scriptFile, delete);
-
-		LOGGER.info("Path del script: " + scriptFile); //$NON-NLS-1$
 		try {
-			Object o;
+			String result;
 			if (administratorMode) {
-				o = script.runAsAdministrator();
+				result = script.runAsAdministrator();
 			}
 			else {
-				o = script.run();
+				result = script.run();
 			}
-			return o;
+			return result;
 		}
 		catch (final IOException e) {
 			throw new IOException("Error en la ejecucion del script via AppleScript: " + e, e); //$NON-NLS-1$
 		}
 	}
 
-	/** Mata el proceso de AutoFirma cuando estamos en macOS. En el resto de sistemas
-	 * no hace nada.
+	/** Mata el proceso intermedio de AutoFirma cuando estamos en macOS. En el resto de sistemas
+	 * no hace nada. El proceso intermedio no es la aplicaci&oacute;n Java, sino el ejecutable que
+	 * traslada los parametros proporcionados por el navegador Web a la aplicaci&oacute;n Java.
 	 * @param processIdText Texto que permite identificar el proceso. */
 	public static void closeMacService(final String processIdText) {
+
+		if (processIdText == null) {
+			LOGGER.warning("No se ha proporcionado el id del proceso que hay que cerrar"); //$NON-NLS-1$
+			return;
+		}
+
 		LOGGER.warning("Ejecuto kill"); //$NON-NLS-1$
 		final AppleScript script = new AppleScript(
 				"kill -9 $(ps -ef | grep " + processIdText + " | awk '{print $2}')"  //$NON-NLS-1$ //$NON-NLS-2$
